% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.R
\name{build_cluster}
\alias{build_cluster}
\title{build clusters}
\usage{
build_cluster(x, centers)
}
\value{
a \code{vector} of indices, indicating the allocated class
}
\description{
build clusters
}
\keyword{internal}
