% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.R
\name{cluster_indices}
\alias{cluster_indices}
\title{choose cluster indices}
\usage{
cluster_indices(cluster, m, similar = TRUE)
}
\arguments{
\item{cluster}{a \code{vector} of indices, indicating the allocated class}

\item{m}{the number of required indices}

\item{similar}{either \code{TRUE} for similar classes, or \code{FALSE} for dissimilar}
}
\value{
a subset of \code{ind} of length \code{m}
}
\description{
choose cluster indices
}
\keyword{internal}
