% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion-package.R
\docType{package}
\name{portion-package}
\alias{portion-package}
\alias{_PACKAGE}
\title{portion: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
