% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion-package.R
\docType{package}
\name{portion-package}
\alias{portion-package}
\alias{_PACKAGE}
\title{portion: Extracting a Data Portion}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Extracting a data portion.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/portion}
  \item Report bugs at \url{https://github.com/loelschlaeger/portion/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com}

}
\keyword{internal}
