% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portion.R
\name{portion}
\alias{portion}
\alias{portion.numeric}
\alias{portion.matrix}
\alias{portion.data.frame}
\alias{portion.list}
\title{Extracting a data portion}
\usage{
portion(x, proportion, how, ...)

portion.numeric(x, proportion, how, ...)

portion.matrix(x, proportion, how, byrow, ignore, cluster, ...)

portion.data.frame(x, proportion, how, byrow, ignore, cluster, ...)

portion.list(x, proportion, how, ...)
}
\arguments{
\item{x}{an object to be portioned}

\item{proportion}{a \code{numeric} between 0 and 1, the relative portion size}

\item{how}{a \code{character}, specifying the portion method, one of:
- \code{"random"} (default), portion at random
- \code{"first"}, portion to the first elements
- \code{"last"}, portion to the last elements
- \code{"similar"}, portion to similar elements based on clustering
- \code{"dissimilar"}, portion to dissimilar elements based on clustering}

\item{...}{further arguments to be passed to or from other methods}

\item{byrow}{\code{TRUE} to portion row-wise or \code{FALSE} to portion column-wise}

\item{ignore}{(relevant if \code{how} is \code{"similar} or \code{"dissimilar)})
an \code{integer} vector of row indices (or column indices if
\code{byrow = FALSE}) to ignore during clustering}

\item{cluster}{(relevant if \code{how} is \code{"similar} or \code{"dissimilar)})
passed on to \code{\link[stats]{kmeans}}}
}
\value{
the portioned input \code{x} with the (row, column, ...) indices used
added as attributes \code{"indices"}
}
\description{
extract a portion of data saved as a \code{vector}, \code{matrix},
\code{data.frame}, or \code{list}
}
